+++ sensor.h
@@ -2,8 +2,9 @@
 #ifndef __sysclean_sensor__
 #define __sysclean_sensor__
 
-#include "utils.h"
 #include <stdint.h>
+#include "msp-cpu.h"
+#include "utils.h"
 
 //volatile unsigned char TIMER_FLAG;
 
@@ -11,6 +12,6 @@
 #define MAX_VALUE 150
 
 double read_sensor_cm();
-double sensor_get();
+float sensor_get();
 
 #endif
